+++ src/constants.js
@@ -1,0 +1,16 @@
+export const tableHeaders = [
+  'Recipe Name',
+  'Description',
+  'Allergens',
+  'Instructions'
+]
+
+export const weekDays = [
+  'Sunday',
+  'Monday',
+  'Tuesday',
+  'Wednesday',
+  'Thursday',
+  'Friday',
+  'Saturday'
+]
